Add bootstrap with spec option
This option seems broken. It makes the initial bootstrap hang for
like 10 minutes and the cephadm.log just shows it waiting to aquire
a lock.

If I ctrl-c the hanging task and just re-run with spec_on_bootstrap
set to false, then 60 seconds later my ansible transaction is done
and the ceph cluster comes up with all resources in <5 minutes.