In standard deploy disable enable-swap.yaml
This was causing a problem for me when I switched to latest tht
and latest tripleo-ansible. I suspect somehting got out of sync
(e.g. puppet on the image). I don't seem to need this right now
so I'm just going to stop uring it so I can workaround it.

The problem was that controller{0,1,2} would get hung and drop
off of the network. I had to reboot them to bring them back.