Make git refresh pull in stacked patch
Patch 2 is based on patch 1.

1. 770674: Introduce tripleo_cephadm role
2. 771034: Introduce tripleo_run_cephadm role and cephadm playbook

Pull in the latest of both but pull in patch 2 after patch 1 so
that I ended up default with patch 2 ready to run. I might have
to manually rebase as I introduce changes, but when I need to
make changes to 1, the latest copy will be ready. Do not delete
the older branch so that I can have both to work on. git review
is smart enough to re-use and update without the need to delete
the old branch.