Add cephadm-extra-vars
As per recent developments [1] we are genereating cephadm-extra-vars
from Heat and tripleo_run_cephadm and passing them to tripleo_cephadm.
This commit includes an example of the genereated ceph-adm-extra-vars
from both heat and ansible and deploy_ceph.sh passes them when it calls
tripleo_cephadm. The cephadm.yaml playbook is updated to be more like
the current one [2]. Also we will have our own ceph_spec.yml as the
current one [3] is for standalone which makes sense at the moment for
tripleo-ci.

[1]
https://review.opendev.org/c/openstack/tripleo-heat-templates/+/767294/25..27/deployment/cephadm/ceph-base.yaml
https://review.opendev.org/c/openstack/tripleo-ansible/+/770674/17..20/tripleo_ansible/roles/tripleo_cephadm/tasks/pools.yaml
https://review.opendev.org/c/openstack/tripleo-ansible/+/771034/18..24/tripleo_ansible/roles/tripleo_run_cephadm/tasks/prepare.yml

[2]
https://review.opendev.org/c/openstack/tripleo-ansible/+/771034/1..24/tripleo_ansible/playbooks/cephadm.yml

[3]
https://review.opendev.org/c/openstack/tripleo-ansible/+/770674/1..20/tripleo_ansible/roles/tripleo_cephadm/files/ceph_spec.yaml