Deploy DCN without generated-container-prepare.yaml
Force TripleO to create the list of containers based on
the list of services found in the roles being deployed.
When I previously tested the CephVolmeEdge service I used
a generated-container-prepare.yaml which already had the
path to the Cinder containers. This hid the need for [1].
Also, if a copy of tripleo-common is in the home dir, use
the latest version of container_image_prepare_defaults.yaml.

[1] https://review.opendev.org/c/openstack/tripleo-common/+/782009